refactor: extract shared frontend components and reduce duplication (#49, #50, #51)
- Extract IndicatorCell to components/indicator-cell.tsx
- Extract HoldingsGrid to components/holdings-grid.tsx (shared by asset-detail and pseudo-etf-detail)
- Extract chart theme/options to lib/chart-utils.ts (shared by 4 chart files)
- Move formatChangePct to lib/format.ts
- Split pseudo-etf-detail.tsx charts into components/chart/pseudo-etf-charts.tsx
- Extract AddConstituentPicker to its own component file
- pseudo-etf-detail.tsx reduced from 746 to 215 lines
